updated make_TP_maps
I included a few extra functions (that already existed) in the code, for example getting the average spectrum over the whole field. I think the mask can be also used here to get a better average spectrum on voxels that there is "real emission".

Now, I have to implement some of the gaussian fitting routines, so I can automateically get the systemic velocity and width of the emission lines.