finished the 2D spectral map code & fixed some minor things in make TP maps
the 2D spectral maps are now working fine. Some things could be improved like, plotting the position of the source rather than velocity and intensity as LABELS.

The modifications to the TP maps is that I added the possibility of giving the molecule name instead of the spectral window for creation of spectral maps.